[ENH] Make regressions use pd.ols
I created two new regression calculations
 - regression
 - rolling_regression

These two functions return the actual ols object so functions
can have access to the other metrics the object contains.

I also added a 'return_ols_obj' kwarg to functions that
call the regression funcs. If it is True, the ols object is
also returned. This is for the same reason as above.