[MAINT] Removes return_ols_model kwargs
In favor of simplicity I removed the option to return
the ols model from the calculations that perform
regressions. e.g. It is simpler if the beta function just
returns the beta coeffecients.

All of the regression calculations still call functions
that return the underlying model as I think it will be
important to be able to access other model data in the future.