[ENH] Add returns function registration.
Here is a really simple implementation of a returns function
registry mechanism.

This doesn't feel like a great way to go about this but it works.
Maybe we can use a class/decorator pattern but I'm not sure what
schools of thought are out there on this sort of thing.

Any thoughts on this, if it's fine I'll document it up and submit a PR