[BUG] Fix timezone compatability issues
There was an issue using timestamp.astimezone(pytz.UTC) that
caused an error depending on whether the timestamp was tz aware
or not. This makes sure that timestamps are normalized to
midnight UTC time.

Note: This could make using intraday data impossible, should we
keep the possibility to make intraday tear sheets? The alternative
would be to convert/localize to UTC, but don't normalize to midnight,
which would not exclude intraday stuff.