BUG Fix caching logic for SPY and Fama French.
Instead of using the timestamp of the file, load the file and look for the last date. This should be more robust.

In addition, if no end-date was passed, we were defaulting to the current time (which will always be more recent than our cache). Instead, this change takes the previous business day (as data will only be available after close on that day, with some delay).