BUG Store SPY with header.
pandas.read_csv() was trying to infer a header from spy.csv which failed as none was present. Now get_symbol_from_yahoo() correctly returns a DataFrame with the column name as the symbol. We then convert this to a Series before we return to the user because the rest of pyfolio assumes benchmark_returns to be a Series.