MAINT Fix gross leverage - round 2
This is separate branch that changes the `long_short_exposure`
function to match gross leverage.

My git repo got out of whack while working on the other branch
so I figured I'd start again

This only changes the code in pyfolio, I did not re-run the notebook
since it seemed to cause an issue last time. I'll submit the notebook
change in a separate PR.