Add stochastic volatility to tear plots
Posterior PDF for log(sigma) and log(nu).  Also plots the stochastic
volatility on top of the returns.

Add text to Bayesian notebook

Explain the 8th and 9th rows in the Bayesian ipython notebook
demonstrating the tear sheet.

Also, a few grammatical edits to the rest of the Bayesian notebook.