BUG: memory leak
The Figures created in the tear sheet functions should be closed when
'return_fig' is False. The closing of Figures should be done after the
they are shown with plt.show() otherwise they are not displayed.


From matplotlib documentation:

The memory required for a Figure is not completely released until the
figure is explicitly closed with close(). Deleting all references to the
figure, and/or using the window manager to kill the window in which the
figure appears on the screen, is not enough, because pyplot maintains
internal references until close() is called